Create utterances.json
Hello, thank you for doing the new FAQs for LogAnalysis - I have created an utterances JSON file in the repo too as this is what is used by Chatbot to onboard this service. Please could you review and approve this content ( if possible by tomorrow as we have a Chatbot build scheduled for them). We would like the utterances.json to be merged and pushed to the prod repo as that is where we build chatbot from - https://github.ibm.com/cloud-doc-build/indexer-publish/tree/output/content/Log-Analysis-with-LogDNAThank you for all your help with this :-) Claire